Ai System + Detection Test
Finish An Ai navmesh system, will follow player base on the shortest calculate distance.

The enemies will have a detection radius, 90 for front and 25 from other side, if tag player was detect in the range, the enemy will chase the player for () second before stopping.

Also include basic health system for testing, if enemies get within the 1.5 radius of player for () second, player will take damage ##this is for animation placeholder